fix: point package entry points to dist/ and consolidate test helpers
Package exports for 6 library packages (core, clickhouse, codegen,
plugin-pull, plugin-typegen, plugin-backfill) were pointing to src/
which breaks npm consumers. Updated main, types, and exports to
dist/. Also set changeset access to public for scoped packages and
merged commands.e2e-helpers into testkit.test.ts.
