+++ packages/cli/src/check.e2e.test.ts
@@ -2,8 +2,7 @@ import { describe, expect, test } from 'bun:test'
 import { rm, writeFile } from 'node:fs/promises'
 import { basename, join } from 'node:path'
 
-import { sortedKeys } from './commands.e2e-helpers'
-import { createFixture, runCli } from './testkit.test'
+import { createFixture, runCli, sortedKeys } from './testkit.test'
 
 describe('@chx/cli check e2e', () => {
   test('check --json fails when migrations are pending', async () => {
